+++ packages/react/src/themes/greyVest/index.js
@@ -37,6 +37,7 @@ import Box from './Box'
 import LinkButton from './LinkButton'
 import TreePauseButton from './TreePauseButton'
 import ToggleFiltersButton from './ToggleFiltersButton'
+import ToggleFiltersHeader from './ToggleFiltersHeader'
 export { default as SearchLayout } from './SearchLayout'
 import BaseSearchFilters from './SearchFilters'
 import DateInput from './DateInput'
@@ -59,6 +60,7 @@ export {
   LinkButton,
   TreePauseButton,
   ToggleFiltersButton,
+  ToggleFiltersHeader,
 }
 
 export let SearchTree = () => {}
@@ -384,7 +386,7 @@ export let GVStyle = () => (
         margin: 5px 0;
       }
       .contexture-facet .gv-checkbox {
-        margin: 0 10px;
+        margin: 0 10px 0 0;
       }
       .contexture-facet > .gv-input[type="text"] {
         margin-bottom: 10px;
@@ -705,7 +707,7 @@ export let Adder = ModalFilterAdder({
 })
 
 export let Modal = defaultProps({ className: 'gv-body' })(DefaultModal)
-
+export let ButtonGroup = defaultProps({ className: 'gv-button-group' })(Flex)
 export let PagerItem = observer(({ active, disabled, ...x }) => (
   <span
     className={`gv-pager-item ${disabled ? 'disabled' : ''} ${
@@ -736,6 +738,7 @@ export let ExampleTypes = ExampleTypeConstructor({
   TagsInput,
   Icon,
   DateInput,
+  ButtonGroup,
 })
 export let Pager = props => (
   <ExampleTypes.ResultPager